+++ NEWS.md
@@ -1,7 +1,8 @@
 # eatGADS 0.15.2.9000
 
 * bug fix `collapseColumns()` (now supports all column names in look up table)
-* `checkValue()` now returns occurence count as vector instead of giving a message
+* `checkValue()` now returns occurrence count as vector instead of giving a message
+* `removeValLabels()` now can remove only specific `value` - `valLabel` pairs if required 
 
 # eatGADS 0.15.2